backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai>=1.0.0
python-docx
pandas
